==> 1_Estimation_varcomp/2_Convergence_Diagnostics.Rmd <==
---
title: "Convergence Diagnostic"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This document contains the code and data to make the Convergence Diagnostic for each model fitted on Estimation of direct and social effects of feeding duration in growing pigs using record from automatic feeding stations.

$Bayesian\ estimating\ of\ variance\ components\ (proportion\ of\ variance)\ with\ Stan\ program,\ on\ visit\ length\ time\ at\ the\ feeder\\ using\ two\ subsets\ of\ data,\ from\ 7\ trials$

The equations for the three mixed model fitted were:

1. $\bf{M1}$:

$$y=X\beta + Zu + e$$ 
Where, $y$ is a $n x 1$ vector  of  visit  length  at the  feeder  (minutes), $X\beta$ are the fixed effects, $location-trial \ (14)$, $hour \ entry \ at \ the \ feeder \ (23)$ and $animal \ median \ weigth$ as covariate; $Z$ is a $n$ x $q$ desing matrix ($q$ is the number of pigs) relates records in $y$ to the random vector of additive genetic effects $u$ ($q$ x $1$); $e$ ($n$ x $1$) is the random residuals vector.

2. $\bf{M2}$:

$$y=X\beta + Zu + Z_fa_f + e$$ 
Fixed, genetic direct and error random effects were same as in equation for $\bf{M1}$, now $Z_f$ is the design matrix of the next individual that visited the feeder, named $followers$, relating to $y$ with the random vector effects $a_f$ ($q$ x $1$).

3. The third model fitted ($\bf{M3}$) included a covariance structure between direct and follower random effects.


$For\ the\ fixed\ parameters\ in\ {\beta},\ flat\ priors\ were\ assumed,\ thus\ {\beta}\ {\sim} U(-\infty,+\infty),\ the\ priors\ distributions\ for\\ covariance\ components\ {\sigma}_d^2,\ {\sigma}_f^2\ {\sim }\ U(0,100),\ {\rho} {\sim}\ U(-1,1)\ and\ the\ prior\ distribution\ for\ the\ error\ variance\\ was\ {\sigma}_e^2\ {\sim}\ U(0,\infty).$

The options to fit the model and to obtain the samples for the parameters were:

$iter = 12000,\ chain= 3,\ burn-in = 2000,\ Thin=1.$


```{r, include=FALSE}
rm(list = ls())
setwd("~/Documents/Github_Visit_Duration/1_Estimation_varcomp")

# 1. Load library and data file

library(tidyverse)
library(coda)
library(mcmcplots)
library(rstan)
library(ggmcmc)
library(bayesplot)

# 2. Load Input files:  class objects: stanfit
# 2.1. R objects obtained from fit models with Immediate replacement subset (IRT)
load("M1_60s_trials.Rdata") # M1
load("M2_60s_trials.Rdata") # M2
load("M3_60s_trials.Rdata") # M3

# 2.2. R objects obtained from fit models with No immediate replacement subset (NIRT)
load("M1_600s_trials.Rdata") # M1
load("M2_600s_trials.Rdata") # M2
load("M3_600s_trials.Rdata") # M3

# 3. Setting options to rstan
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = TRUE)

```

## 1. Immediate replacement time (IRT) subset

### 1.1. Mixed model: M1
```{r, include=FALSE}
outp1 <- As.mcmc.list(M160s.model)

for (i in 1:length(outp1)) {
  colnames(outp1[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),
                          "Median Weight",
                          "var_eartag","var_error",
                          "prp_var_eartag","prp_var_error", "lp__")}
# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")

```
#### 1.1.1. Check autorrelation, effective sample size, traceplot
```{r, warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%c("Median Weight",
  "var_eartag","var_error","prp_var_eartag","prp_var_error")))

autocorr.diag(outp1)
effectiveSize(outp1)
geweke.diag(outp1)
gelman.diag(outp1, transform = T)
traplot(outp1,col =c("red1","blue4","purple3"))
denplot(outp1)
```

#### 1.1.2. Summary Posterior Distribution
```{r}
summary(outp1)
print(M160s.model)
```
#### 1.1.3. Posterior Correlation of model parameters
```{r, warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp1,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```

### 1.2. Mixed model: M2
```{r, include=FALSE}
outp2 <- As.mcmc.list(M260s.model)

for (i in 1:length(outp2)) {
  colnames(outp2[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),"Median Weight",
                          "var_eartag", "var_follower", "var_error", 
                          "prp_var_eartag","prp_var_follower", "prp_var_error","lp__")}

# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")
```

#### 1.2.1.Check autorrelation, effective sample size, traceplot, density plot.
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%c("Median Weight",
                                                      "var_eartag", "var_follower", "var_error",
                                                      "prp_var_eartag","prp_var_follower", "prp_var_error")))



autocorr.diag(outp2)
effectiveSize(outp2)
geweke.diag(outp2)
gelman.diag(outp2, transform = T)
traplot(outp2, col = c("red1","purple1","blue4"))
denplot(outp2,col = c("red1","purple1","blue4"))

```

#### 1.2.2. Summary Posterior Distribution
```{r}
summary(outp2)
print(M260s.model)
```

#### 1.2.3. Posterior Correlation of model parameters
```{r, warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp2,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```

## 1.3. Mixed model: M3
```{r, include=FALSE}
outp3<-As.mcmc.list(M360s.model)

for (i in 1:length(outp3)) {
  colnames(outp3[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),"Median Weight",
                          "rho","var_eartag","var_follower", "var_error", 
                          "prp_var_eartag","prp_var_follower", "prp_var_error","lp__")}

# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")

```

#### 1.3.1. Check autorrelation, effective sample size, traceplot, density plot.
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%c("Median Weight",
                                                       "rho","var_eartag","var_follower", "var_error", 
                                                       "prp_var_eartag","prp_var_follower", "prp_var_error")))

autocorr.diag(outp3)
effectiveSize(outp3)
geweke.diag(outp3)
gelman.diag(outp3, transform = T)
traplot(outp3,col =c("red1","blue4","purple3"))
denplot(outp3,col = c("red1","blue4","purple3"))
```

#### 1.3.2. Summary Posterior Distribution
```{r}
summary(outp3)
print(M360s.model)
```

#### 1.3.3. Posterior Correlation of model parameters
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp3,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```


## 2. No immediate replacement time (NIRT) subset

### 2.1. Mixed model: M1
```{r, include=FALSE}
outp1 <- As.mcmc.list(M1600s.model)

for (i in 1:length(outp1)) {
  colnames(outp1[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),
                          "Median Weight",
                          "var_eartag","var_error","prp_var_eartag","prp_var_error","lp__")}
# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")
```

#### 2.1.1.Check autorrelation, effective sample size, traceplot, density plot.
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp1)%>%filter(Parameter%in%c("Median Weight",
                                                       "var_eartag","var_error","prp_var_eartag" ,"prp_var_error")))

autocorr.diag(outp1)
effectiveSize(outp1)
geweke.diag(outp1)
gelman.diag(outp1,transform = T,multivariate = F)
traplot(outp1,col =c("red1","blue4","purple3"))
denplot(outp1)
```

#### 2.1.2. Summary Posterior Distribution
```{r}
summary(outp1)
print(M1600s.model)
```

#### 2.1.3. Posterior Correlation of model parameters
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp1,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```

### 2.2. Mixed model: M2
```{r, include=FALSE}
outp2 <- As.mcmc.list(M2600s.model)

for (i in 1:length(outp2)) {
  colnames(outp2[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),
                          "Median Weight","var_eartag","var_follower", "var_error","prp_var_eartag","prp_var_follower",  "prp_var_error","lp__")}

# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")
```

#### 2.2.1.Check autorrelation, effective sample size, traceplot, density plot.
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp2)%>%filter(Parameter%in%c("Median Weight",
                                                       "var_eartag","var_follower", "var_error","prp_var_eartag","prp_var_follower",  "prp_var_error")))

autocorr.diag(outp2)
effectiveSize(outp2)
geweke.diag(outp2)
gelman.diag(outp2, transform = T, multivariate = F)
traplot(outp2, col = c("red1","purple1","blue4"))
denplot(outp2,col = c("red1","purple1","blue4"))
```

#### 2.2.3. Summary Posterior Distribution
```{r}
summary(outp2)
print(M2600s.model)
```

#### 2.2.3.Posterior Correlation of model parameters
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp2,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```


### 2.3. Mixed model: M3
```{r,include=FALSE}

outp3<-As.mcmc.list(M3600s.model)

for (i in 1:length(outp3)) {
  colnames(outp3[[i]])<-c(paste("L1_t",seq(1,7), sep = "_"),
                          paste("L2_t",seq(1,7), sep = "_"),
                          paste("h",seq(1,23), sep = "_"),
                          "Median Weight", "rho","var_eartag","var_follower", "var_error", "prp_var_eartag", "prp_var_follower", "prp_var_error","lp__")}

# autocorrelation plot by chain, 
lt1<-paste("L1_t",seq(1,7), sep = "_")
lt2<-paste("L2_t",seq(1,7), sep = "_")
hent1<-paste("h",seq(1,12), sep = "_")
hent2<-paste("h",seq(13,23), sep = "_")
```

#### 2.3.1. Check autorrelation, effective sample size, traceplot, density plot
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%lt1))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%lt2))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%hent1))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%hent2))
ggs_autocorrelation(ggs(outp3)%>%filter(Parameter%in%c("Median Weight",
                                                       "rho","var_eartag","var_follower", "var_error", 
                                                       "prp_var_eartag", "prp_var_follower", "prp_var_error")))

autocorr.diag(outp3)
effectiveSize(outp3)
geweke.diag(outp3)
gelman.diag(outp3, transform = T,multivariate = F)
traplot(outp3,col =c("red1","blue4","purple3"))
denplot(outp3,col = c("red1","blue4","purple3"))
```

#### 2.3.2. Summary Posterior Distribution
```{r}
summary(outp3)
print(M3600s.model)
```

#### 2.3.3. Posterior Correlation of model parameters
```{r,warning=FALSE, message=FALSE,fig.dim=c(10,8)}
parcorplot(outp3,col = terrain.colors(15,0.5,T), cex.axis=0.6)
```
